% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conpower.R
\name{survival}
\alias{survival}
\title{calculates conditional power for Time-to-event Endpoint}
\usage{
survival(ek, E, observedHR, se_survival, futureHR, zfinalsurv)
}
\arguments{
\item{ek}{Number of events obtained at interim stage/look}

\item{E}{Number of events at final stage/look}

\item{observedHR}{observedHR Observed Hazard Ratio at Interim stage/look}

\item{se_survival}{Standard Error of log(Hazard Ratio)}

\item{futureHR}{Assumed Hazard Ratio for Future Patients}

\item{zfinalsurv}{Final stage/Look Z-Statistic value which was obtained while planning analysis in Design Stage}
}
\value{
conditional power for time-to-event endpoint
}
\description{
This package will calculate conditional power for stopping a trial for futility or stop for efficacy. It is possible to calculate the conditional power of the study to reject the null hypothesis given the current results obtained from gsDesign (Group sequential designs). An user has to first calculate the sample size for any types of endpoint: Continuous, Binary and Time-to-Event based on R-package gsDesign. Then using this results one can calculate the Conditional power, which is one of the tools, when computed over a range of alternatives, can be of guidance in
deciding whether to continue the study given those available information.
}
\examples{
survival(ek=300, E=377, observedHR=0.75, se_survival=0.164, futureHR=0.80, zfinalsurv=2.16)

}
\references{
Jennison, C. and Turnbull, B.W. (2005). Group Sequential Methods with Applications to Clinical Trials. Boca Raton: Chapman and Hall.

Chang, M. (2015). Introductory Adaptive Trial Designs: A Practical Guide with R. CRC press, Chapman and Hall.
}
\author{
Mohammad Anamul Haque, Tomasz Burzykowski, Emmanuel Quinaux and Nahid Sultana
}
